backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
sqlparse==0.5.5
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
sqlparse==0.5.5
